project 3/requirements.txt
adbc-driver-manager==1.2.0
adbc-driver-postgresql==1.2.0
aiobotocore==2.14.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
botocore==1.35.7
certifi==2024.8.30
cffi==1.17.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.4
jmespath==1.0.1
minio==7.2.8
multidict==6.0.5
mysql-connector-python==9.0.0
numpy==2.1.1
pandas==2.2.2
polars==1.6.0
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3fs==2024.9.0
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.10.0